refactor: bugs when saving utc in supabase
