+++ app/[idTeam]/meetings/CreateMeeting.tsx
@@ -3,7 +3,6 @@
 // Create a variable for all possibles UTC timezones in Earth with their respective offsets
 import { utcTimezones } from "@/lib/types";
 
-import { useToast } from "@/hooks/use-toast";
 import { useTeamMeeting } from "@/hooks/useTeamMeeting";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
@@ -82,7 +81,6 @@ export const CreateMeeting = ({
   const { createTeamMeeting } = useTeamMeeting();
   const [isDialogOpen, setIsDialogOpen] = useState(false);
   const [savedZones, setSavedZones] = useState<any[]>([]);
-  const { toast } = useToast();
 
   const combineDateAndTime = (
     date: Date | null,
@@ -146,10 +144,6 @@ export const CreateMeeting = ({
         form.reset();
         onMeetingCreated();
         setIsDialogOpen(false);
-        toast({
-          title: "Meeting created",
-          description: "Your meeting has been created successfully.",
-        });
       } else {
         console.error("Failed to create meeting.");
       }
